moved table related components
