+++ frontend/src/components/managePermissions/ConfirmDialogText.vue
