+++ frontend/src/components/managePermissions/ManagePermissions.vue
@@ -4,7 +4,8 @@ import Dropdown, { type DropdownChangeEvent } from 'primevue/dropdown';
 import TabView from 'primevue/tabview';
 import TabPanel from 'primevue/tabpanel';
 import ManagePermissionsTitle from '@/components/managePermissions/ManagePermissionsTitle.vue';
-import UserDataTable from '@/components/managePermissions/UserDataTable.vue';
+import UserDataTable from '@/components/managePermissions/table/UserDataTable.vue';
+import ApplicationAdminTable from './table/ApplicationAdminTable.vue';
 
 import {
     applicationsUserAdministers,
